crawler/requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.5.0
pydantic==2.7.1
python-dateutil==2.9.0
pytest==8.2.0
pytest-asyncio==0.23.7
